Move Web Context Menu into its own class (#534)
